+++ src/include/_c_interpolation_catmull_rom.h
@@ -3,18 +3,9 @@
 
 // Cubic function used in Catmull-Rom interpolation
 // https://en.wikipedia.org/wiki/Cubic_Hermite_spline#Catmull.E2.80.93Rom_spline
-double _c_cubic(double v) {
-  double a = 0.5;
-  double z = 0;
-  if (v < 0) {
-    v = -v;
-  }
-  if (v < 1) {
-    z = v * v * (v * (-a + 2) + (a - 3)) + 1;
-  } else if (v < 2) {
-    z = -a * v * v * v + 5 * a * v * v - 8 * a * v + 4 * a;
-  }
-  return z;
-}
+double _c_cubic(double v);
+
+// Catmull-Rom interpolation
+double _c_interpolate(float* image, double r, double c, int rows, int cols);
 
 #endif  // _C_INTERPOLATION_CATMULL_ROM_H
