+++ src/nanopyx/core/transform/interpolation_catmull_rom.pxd
@@ -1,12 +1,15 @@
 cdef extern from "_c_interpolation_catmull_rom.h":
     double _c_cubic(double v) nogil
+    double _c_interpolate(float * image, double r, double c, int rows, int cols) nogil
 
-from .interpolation_nearest_neighbor cimport Interpolator as InterpolatorNearestNeighbor
+
+from .interpolation_nearest_neighbor cimport \
+    Interpolator as InterpolatorNearestNeighbor
 
 # pyx2pxd: starting point
 # Code below is autogenerated by pyx2pxd - https://github.com/HenriquesLab/pyx2pxd
 
-cdef double _interpolate(float[:,:] image, double x, double y) nogil
+cdef double _interpolate(np.ndarray[np.float32_t, ndim=2]  image, double x, double y) nogil
 
 cdef class Interpolator(InterpolatorNearestNeighbor):
     cdef float _interpolate(self, float x, float y) nogil
