+++ src/nanopyx/core/transform/interpolation_catmull_rom.pyx
@@ -1,7 +1,9 @@
 # cython: infer_types=True, wraparound=False, nonecheck=False, boundscheck=False, cdivision=True, language_level=3, profile=False, autogen_pxd=True
 
 import numpy as np
+
 cimport numpy as np
+from cython cimport view
 
 
 def interpolate(np.ndarray im, double x, double y) -> float:
@@ -23,40 +25,45 @@ cdef double _interpolate(float[:,:] image, double x, double y) nogil:
     :param y: y-coordinate to interpolate at
     :return: Interpolated pixel value (float)
     """
-    cdef int w = image.shape[1]
-    cdef int h = image.shape[0]
+    cdef int rows = image.shape[0]
+    cdef int cols = image.shape[1]
 
-    # return 0 if x OR y positions do not exist in image
-    if not 0 <= x < w or not 0 <= y < h:
-        return 0
+    # # return 0 if x OR y positions do not exist in image
+    # if not 0 <= x < w or not 0 <= y < h:
+    #     return 0
 
-    cdef int x0 = int(x)
-    cdef int y0 = int(y)
+    # cdef int x0 = int(x)
+    # cdef int y0 = int(y)
 
-    # do not interpolate if x and y positions exist in image
-    #if x == x0 and y == y0:
-    #    return image[y0, x0]
+    # # do not interpolate if x and y positions exist in image
+    # #if x == x0 and y == y0:
+    # #    return image[y0, x0]
 
-    cdef int u0 = int(x - 0.5)
-    cdef int v0 = int(y - 0.5)
-    cdef double q = 0
-    cdef double p
-    cdef int v, u, i, j, _u, _v
+    # cdef int u0 = int(x - 0.5)
+    # cdef int v0 = int(y - 0.5)
+    # cdef double q = 0
+    # cdef double p
+    # cdef int v, u, i, j, _u, _v
 
-    for j in range(4):
-        v = v0 - 1 + j
-        p = 0
-        for i in range(4):
-            u = u0 - 1 + i
-            _u = max(0, min(u, w-1))
-            _v = max(0, min(v, h-1))
-            p = p + image[_v, _u] * _c_cubic(x - (u + 0.5))
-        q = q + p * _c_cubic(y - (v + 0.5))
+    # for j in range(4):
+    #     v = v0 - 1 + j
+    #     p = 0
+    #     for i in range(4):
+    #         u = u0 - 1 + i
+    #         _u = max(0, min(u, w-1))
+    #         _v = max(0, min(v, h-1))
+    #         p = p + image[_v, _u] * _c_cubic(x - (u + 0.5))
+    #     q = q + p * _c_cubic(y - (v + 0.5))
 
-    #if isnan(q) or isinf(q):
-    #    return 0.
+    # #if isnan(q) or isinf(q):
+    # #    return 0.
 
-    return float(q)
+    # return float(q)
+
+    # Get the data pointer of the numpy array
+
+    # Call the C function with the numpy array's data pointer
+    return _c_interpolate(&image[0], y, x, rows, cols)
 
 
 cdef class Interpolator(InterpolatorNearestNeighbor):
